[DHS-1678]: fix progress bar error for files over 2^31 bytes
Internally, iRODS rule language's int() calls C's atoi() which will fail
for integers over (signed) INT_MAX (~2^31). So the AVU now is set to a
string representing a float instead of an integer.

+++ misc/policies.r
@@ -11,10 +11,10 @@ acPostProcForPut {
         if( *resource == $rescName ) {
             foreach (*av in SELECT META_COLL_ATTR_NAME, META_COLL_ATTR_VALUE WHERE COLL_NAME == "/nlmumc/projects/*project/*collection") {
                 if ( *av.META_COLL_ATTR_NAME == "sizeIngested" ) {
-                    *sizeIngested = int(*av.META_COLL_ATTR_VALUE);
+                    *sizeIngested = double(*av.META_COLL_ATTR_VALUE);
                 }
             }
-            *sizeIngested = *sizeIngested + int($dataSize);
+            *sizeIngested = *sizeIngested + double($dataSize);
             msiAddKeyVal(*metaKV,  'sizeIngested', str(*sizeIngested));
             msiSetKeyValuePairsToObj(*metaKV, "/nlmumc/projects/*project/*collection", "-C");
         }
